Refactor: Move dimension to common internal package
