+++ src/main/kotlin/com/amazon/opendistroforelasticsearch/indexmanagement/common/model/dimension/DateHistogram.kt
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.indexmanagement.rollup.model.dimension
+package com.amazon.opendistroforelasticsearch.indexmanagement.common.model.dimension
 
 import org.elasticsearch.common.io.stream.StreamInput
 import org.elasticsearch.common.io.stream.StreamOutput
@@ -73,11 +73,12 @@ data class DateHistogram(
         out.writeZoneId(timezone)
     }
 
-    override fun toSourceBuilder(): CompositeValuesSourceBuilder<*> {
+    override fun toSourceBuilder(appendType: Boolean): CompositeValuesSourceBuilder<*> {
         val calendarInterval = this.calendarInterval
         val fixedInterval = this.fixedInterval
+        val name = if (appendType) "${this.targetField}.${Type.DATE_HISTOGRAM.type}" else this.targetField
 
-        return DateHistogramValuesSourceBuilder(this.targetField)
+        return DateHistogramValuesSourceBuilder(name)
             .field(this.sourceField)
             .timeZone(this.timezone)
             .apply {
