+++ src/main/kotlin/com/amazon/opendistroforelasticsearch/indexmanagement/common/model/dimension/Terms.kt
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.indexmanagement.rollup.model.dimension
+package com.amazon.opendistroforelasticsearch.indexmanagement.common.model.dimension
 
 import org.elasticsearch.common.io.stream.StreamInput
 import org.elasticsearch.common.io.stream.StreamOutput
@@ -57,8 +57,9 @@ data class Terms(
         out.writeString(targetField)
     }
 
-    override fun toSourceBuilder(): CompositeValuesSourceBuilder<*> {
-        return TermsValuesSourceBuilder(this.targetField)
+    override fun toSourceBuilder(appendType: Boolean): CompositeValuesSourceBuilder<*> {
+        val name = if (appendType) "${this.targetField}.${Type.TERMS.type}" else this.targetField
+        return TermsValuesSourceBuilder(name)
             .missingBucket(true)
             .field(this.sourceField)
     }
