+++ src/main/kotlin/com/amazon/opendistroforelasticsearch/indexmanagement/rollup/util/RollupUtils.kt
@@ -22,10 +22,10 @@ import com.amazon.opendistroforelasticsearch.indexmanagement.rollup.model.Rollup
 import com.amazon.opendistroforelasticsearch.indexmanagement.rollup.model.RollupFieldMapping
 
 import com.amazon.opendistroforelasticsearch.indexmanagement.rollup.model.RollupMetadata
-import com.amazon.opendistroforelasticsearch.indexmanagement.rollup.model.dimension.DateHistogram
-import com.amazon.opendistroforelasticsearch.indexmanagement.rollup.model.dimension.Dimension
-import com.amazon.opendistroforelasticsearch.indexmanagement.rollup.model.dimension.Histogram
-import com.amazon.opendistroforelasticsearch.indexmanagement.rollup.model.dimension.Terms
+import com.amazon.opendistroforelasticsearch.indexmanagement.common.model.dimension.DateHistogram
+import com.amazon.opendistroforelasticsearch.indexmanagement.common.model.dimension.Dimension
+import com.amazon.opendistroforelasticsearch.indexmanagement.common.model.dimension.Histogram
+import com.amazon.opendistroforelasticsearch.indexmanagement.common.model.dimension.Terms
 import com.amazon.opendistroforelasticsearch.indexmanagement.rollup.model.metric.Average
 import com.amazon.opendistroforelasticsearch.indexmanagement.rollup.model.metric.Max
 import com.amazon.opendistroforelasticsearch.indexmanagement.rollup.model.metric.Min
@@ -56,11 +56,7 @@ import org.elasticsearch.search.aggregations.AggregationBuilder
 import org.elasticsearch.search.aggregations.AggregatorFactories
 import org.elasticsearch.search.aggregations.bucket.composite.CompositeAggregationBuilder
 import org.elasticsearch.search.aggregations.bucket.composite.CompositeValuesSourceBuilder
-import org.elasticsearch.search.aggregations.bucket.composite.DateHistogramValuesSourceBuilder
-import org.elasticsearch.search.aggregations.bucket.composite.HistogramValuesSourceBuilder
-import org.elasticsearch.search.aggregations.bucket.composite.TermsValuesSourceBuilder
 import org.elasticsearch.search.aggregations.bucket.histogram.DateHistogramAggregationBuilder
-import org.elasticsearch.search.aggregations.bucket.histogram.DateHistogramInterval
 import org.elasticsearch.search.aggregations.bucket.histogram.HistogramAggregationBuilder
 import org.elasticsearch.search.aggregations.bucket.terms.TermsAggregationBuilder
 import org.elasticsearch.search.aggregations.metrics.AvgAggregationBuilder
@@ -95,39 +91,7 @@ fun Rollup.getRollupSearchRequest(metadata: RollupMetadata): SearchRequest {
 @Suppress("ComplexMethod", "NestedBlockDepth")
 fun Rollup.getCompositeAggregationBuilder(afterKey: Map<String, Any>?): CompositeAggregationBuilder {
     val sources = mutableListOf<CompositeValuesSourceBuilder<*>>()
-    this.dimensions.forEach { dimension ->
-        when (dimension) {
-            is DateHistogram -> {
-                DateHistogramValuesSourceBuilder(dimension.targetField + ".date_histogram")
-                    .missingBucket(true) // TODO: Should this always be true or be user-defined?
-                    .apply {
-                        this.field(dimension.sourceField)
-                        this.timeZone(dimension.timezone)
-                        dimension.calendarInterval?.let { it ->
-                            this.calendarInterval(DateHistogramInterval(it))
-                        }
-                        dimension.fixedInterval?.let { it ->
-                            this.fixedInterval(DateHistogramInterval(it))
-                        }
-                    }.also { sources.add(it) }
-            }
-            is Terms -> {
-                TermsValuesSourceBuilder(dimension.targetField + ".terms")
-                    .missingBucket(true)
-                    .apply {
-                        this.field(dimension.sourceField)
-                    }.also { sources.add(it) }
-            }
-            is Histogram -> {
-                HistogramValuesSourceBuilder(dimension.targetField + ".histogram")
-                    .missingBucket(true)
-                    .apply {
-                        this.field(dimension.sourceField)
-                        this.interval(dimension.interval)
-                    }.also { sources.add(it) }
-            }
-        }
-    }
+    this.dimensions.forEach { dimension -> sources.add(dimension.toSourceBuilder(appendType = true)) }
     return CompositeAggregationBuilder(this.id, sources).size(this.pageSize).also { compositeAgg ->
         afterKey?.let { compositeAgg.aggregateAfter(it) }
         this.metrics.forEach { metric ->
