+++ src/test/kotlin/com/amazon/opendistroforelasticsearch/indexmanagement/transform/TransformRestTestCase.kt
@@ -3,7 +3,7 @@ package com.amazon.opendistroforelasticsearch.indexmanagement.transform
 import com.amazon.opendistroforelasticsearch.indexmanagement.IndexManagementPlugin.Companion.TRANSFORM_BASE_URI
 import com.amazon.opendistroforelasticsearch.indexmanagement.IndexManagementRestTestCase
 import com.amazon.opendistroforelasticsearch.indexmanagement.makeRequest
-import com.amazon.opendistroforelasticsearch.indexmanagement.rollup.model.dimension.Dimension
+import com.amazon.opendistroforelasticsearch.indexmanagement.common.model.dimension.Dimension
 import com.amazon.opendistroforelasticsearch.indexmanagement.transform.model.Transform
 import org.apache.http.HttpEntity
 import org.apache.http.entity.ContentType.APPLICATION_JSON
